+++ server/app/Providers/MailConfigServiceProvider.php
@@ -13,20 +13,16 @@ class MailConfigServiceProvider extends ServiceProvider
 {
     /**
      * Bootstrap the application services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         //
     }
 
     /**
      * Register the application services.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         if (!File::exists(base_path('.env'))) {
             return;