+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes 
 
-## [Unreleased]
+## [1.1.0] - 2026-02-03
 
 ### Added